fix implicit type convert

+++ include/utopia/portal_server/client_connection/tls/srp_helper_functions/tls_prf_sha256.hpp
@@ -12,7 +12,7 @@ namespace utopia::portal::client_connection {
 std::optional<std::vector<std::uint8_t>>
 tls_prf_sha256(const std::vector<uint8_t> &secret,
                const std::vector<uint8_t> &label,
-               const std::vector<uint8_t> &seed, std::uint32_t output_size) {
+               const std::vector<uint8_t> &seed, std::size_t output_size) {
 
   std::vector<std::uint8_t> output(output_size, 0);
